+++ data/substack_newsletters.json
@@ -1,1 +1,6 @@
-{"https://www.noahpinion.blog/": "Noahpinion"}
+[{"https://www.noahpinion.blog/": "Noahpinion"},
+{"https://www.oneusefulthing.org/": "Ethan Mollick's substack"},
+{"https://www.natesilver.net/": "Nate Silver's substack"},
+{"https://www.slowboring.com/": "Matt Yglesias' substack"},
+{"https://semianalysis.substack.com/": "SemiAnalysis (Dylan Patel)"}
+]
