+++ data/youtube_channels.json
@@ -1,1 +1,5 @@
-{"https://www.youtube.com/@EzraKleinShow": "Favorite Podcast"}
+[{"https://www.youtube.com/@EzraKleinShow": "Ezra Klein show (Favorite Podcast)"},
+{"https://www.youtube.com/@DwarkeshPatel": "Dwarkesh Patel (AI Podcast)"},
+{"https://www.youtube.com/@restispolitics": "The Rest Is Politics"},
+{"https://www.youtube.com/@Asianometry": "Asianometry"},
+{"https://www.youtube.com/@LeadingTRIP": "Leading: The Rest Is Politics Podcast"}]
