+++ src/agents/books_agent.py
@@ -1,8 +1,8 @@
 from pathlib import Path
 import sys
 
-sys.path.append(str(Path(__file__).parent.parent.parent))
-from src.tools.home_agent_tools import (
+sys.path.append(str(Path(__file__).parent.parent))
+from src.tools.books_and_movies_tools import (
     add_book_to_reading_list,
     get_book_lists,
     mark_book_read,
