+++ src/agents/home_agent.py
@@ -1,0 +1,30 @@
+from pathlib import Path
+import sys
+
+sys.path.append(str(Path(__file__).parent.parent.parent))
+from src.agents.agent import AgentAbstract
+from src.tools.home_agent_tools import (
+    book_agent_query,
+    movies_agent_query,
+    recommender_agent_query,
+)
+from src.utils.prompts import HOME_AGENT_PROMPT
+
+
+class HomeAgent(AgentAbstract):
+    """
+    HomeAgent is responsible for managing home-related tasks such as querying book and movie agents,
+    and handling recommendations.
+    """
+
+    def get_agent_config(self):
+        """Return the Task agent configuration."""
+        return self.config.books_agent
+
+    def get_prompt(self):
+        """Return the system prompt for this agent."""
+        return HOME_AGENT_PROMPT
+
+    def get_tools(self):
+        """Return the list of tools available for this agent."""
+        return [book_agent_query, movies_agent_query, recommender_agent_query]
