+++ src/pipelines/running_home_agent.py
@@ -1,0 +1,14 @@
+from pathlib import Path
+import sys
+
+sys.path.append(str(Path(__file__).parent.parent.parent))
+from src.agents.home_agent import HomeAgent
+
+
+def main():
+    home_agent = HomeAgent()
+    home_agent.query("What are the youtube channels I am monitoring?")
+
+
+if __name__ == "__main__":
+    main()
